mher: bash script had a bug where it tried to move a not-existing sysmap.h file. Fixed by checking if it exists.

+++ assignment01/gensysmap.sh
@@ -1,6 +1,10 @@
 #! /bin/bash
 
-mv sysmap.h sysmap.h.old
-cat /boot/System.map-`uname -r` |
-grep -P "\s+[RDT]+\s" |
-sed 's/^\([^ ]*\) \([^ ]*\) \([^ ]*\)$/#define \3 \1/g' >>sysmap.h
+# check if a file by the name sysmap.h already exists and back it up
+# if necessary
+[ -f ./sysmap.h ] && mv sysmap.h sysmap.h.old
+
+
+cat /boot/System.map-`uname -r` | 
+	grep -P "\s+[RDT]+\s" |
+	sed 's/^\([^ ]*\) \([^ ]*\) \([^ ]*\)$/#define \3 \1/g' >>sysmap.h
